requirements.txt
geopy
timezonefinder
requests
pytz
python-dotenv
tk
